chore: move PersonalDetails component
