+++ src/Components/Creature/PersonalDetails/PersonalBlock.tsx
@@ -2,7 +2,7 @@ import { TFunction } from 'i18next';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import { Race } from '@wowfinder/model';
-import type { CharXpProps } from '../../types';
+import type { CharXpProps } from '../../CharacterSheet/Page/types';
 // TODO #459: Add support for drop-downs
 import PersonalEntryNumber, {
     PersonalEntryBigNumber,
