+++ src/Components/Creature/PersonalDetails/PersonalEntryNumber.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { fThousands } from '@wowfinder/ts-utils';
-import { borderless, bottomLine } from '../../../../../styles';
+import { borderless, bottomLine } from '../../../styles';
 import { PersonalEntry, PersonalItemProps } from './base';
 
 const StyledInput = styled.input`
