+++ src/Components/Creature/PersonalDetails/index.ts
